pathnames: widen total_str_bytes to size_t for slab sizing
list_to_index() walks the namelist twice: first to sum the per-name
strlen()+1 bytes into total_str_bytes, then to memcpy each name into a
single alloc_shared() slab sized by that sum. The accumulator was an
unsigned int while its sibling findex_bytes is a size_t computed via
shared_size_mul() with an explicit overflow check.

On a 64-bit host this is a latent narrowing: if MAX_WALKED_PATHS (or
the per-name length cap) is ever raised so the sum exceeds UINT_MAX,
the unsigned-int accumulator silently wraps and alloc_shared() returns
a slab that is up to 4 GiB smaller than the strings about to be copied
into it. The second pass then memcpy's past the end of the slab.

Widen total_str_bytes to size_t so it mirrors findex_bytes and matches
the size_t parameter of alloc_shared(). No call sites consume the
value (no format specifiers, no signed compares), so widening is a
mechanical declaration change. Hardening only -- no-op at runtime
under current limits, where the sum cannot reach UINT_MAX.

+++ pathnames.c
@@ -464,7 +464,7 @@ static const char ** list_to_index(struct namelist *namelist)
 	struct namelist *nl;
 	const char **findex;
 	unsigned int i = 0;
-	unsigned int total_str_bytes = 0;
+	size_t total_str_bytes = 0;
 	char *slab;
 	unsigned int slab_off = 0;
